🏷️ Add py.typed marker for type checking support (#62)
- Create empty py.typed file in youtrack_cli/ directory
- Update pyproject.toml to include py.typed in package distribution
- Enables type hints for package users per PEP 561

Fixes #49
