+++ youtrack_cli/commands/__init__.py
@@ -1,0 +1,17 @@
+"""Command modules for YouTrack CLI."""
+
+from .articles import articles
+from .boards import boards
+from .issues import issues
+from .projects import projects
+from .time_tracking import time
+from .users import users
+
+__all__ = [
+    "articles",
+    "boards",
+    "issues",
+    "projects",
+    "time",
+    "users",
+]
